% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaming_descriptors.R
\name{descriptor_terms}
\alias{descriptor_terms}
\title{Access the data dictionary associated with Teaming Core Collaboration Tags}
\usage{
descriptor_terms()
}
\value{
class \code{descriptor_terms} data.frame
}
\description{
Available \href{https://docs.google.com/spreadsheets/d/1wmF-ezE8yAAvIrXrpiHsxwmWC9Oee1LJSotVJsFUSL4/edit#gid=989001416}{here:}
}
\examples{
dterms = descriptor_terms()

head(dterms)

class(dterms)



}
\seealso{
Other data_import: 
\code{\link{dssrc_dataframe}()}
}
\author{
seandavi@gmail.com
}
\concept{data_import}
