% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaming_descriptors.R
\name{descriptor_terms_as_igraph}
\alias{descriptor_terms_as_igraph}
\title{convert descriptor terms to igraph object}
\usage{
descriptor_terms_as_igraph(obj)
}
\arguments{
\item{obj}{a \code{descriptor_terms} data.frame (usually from \code{\link{descriptor_terms}})}
}
\value{
an \code{igraph} object
}
\description{
convert descriptor terms to igraph object
}
\examples{
dterms = descriptor_terms()

g = descriptor_terms_as_igraph(dterms)

g

if(require(visNetwork)) {
  visNetwork::visIgraph(g)
}


}
\concept{visualization}
