% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_work_resource.R
\name{.extract_simple_details_from_work_record}
\alias{.extract_simple_details_from_work_record}
\title{Extract a simple list from work-summary record}
\usage{
.extract_simple_details_from_work_record(work)
}
\description{
This internal helper function extracts the title
and external ids (pmid, pmc, and doi) where available.
}
\seealso{
Other orcid: 
\code{\link{orcid_work_detail}()},
\code{\link{orcid_works_json}()},
\code{\link{works_from_orcid}()}

Other bibliography: 
\code{\link{orcid_work_detail}()},
\code{\link{orcid_works_json}()},
\code{\link{works_from_orcid}()}
}
\concept{bibliography}
\concept{orcid}
\keyword{Internal}
