% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssrc_dataframe.R
\name{dssrc_dataframe}
\alias{dssrc_dataframe}
\title{Get dsSrc data from google sheet}
\usage{
dssrc_dataframe()
}
\description{
The dsSrc dataset is maintained by Julie
and contains most of the relevant onboarding
information.
}
\examples{
dssrc = dssrc_dataframe()
colnames(dssrc)
dplyr::glimpse(dssrc)

}
