% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_key}
\alias{has_key}
\title{Test whether an object with names has specified key}
\usage{
has_key(l, key)
}
\description{
Test whether a list, data.frame(columns), or a named
vector has a specified key in it.
}
\examples{
l = list(a=1,b=2,c=3)
has_key(l, 'b') #TRUE
has_key(l, 'not_present') #FALSE


}
\seealso{
Other utilities: 
\code{\link{safe_get}()}
}
\concept{utilities}
\keyword{Internal}
