% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_api.R
\name{orcid_api_call}
\alias{orcid_api_call}
\title{Helper function for calling orcid API}
\usage{
orcid_api_call(orcid, path = "/works")
}
\arguments{
\item{orcid}{character(1) the required orcid}

\item{works}{character(1) the rest of the API path after the \code{orcid} part.}
}
\value{
an R \code{list} representation of the REST API json response
}
\description{
The ORCID REST API does not require any authentication for basic
querying. The OpenAPI documentation is available.
}
\details{
The API typically works with the ORCID as the base resource. Therefore,
an ORCID is required for this function. The path follows the documentation.
}
\examples{
orcid='0000-0001-7443-925X'

orcid_api_call(orcid)

}
