% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_work_resource.R
\name{orcid_work_detail}
\alias{orcid_work_detail}
\title{Get a ORCID \code{work} detail}
\usage{
orcid_work_detail(orcid, put_code)
}
\arguments{
\item{orcid}{character(1) an orcid is required}

\item{put_code}{integer(1) the put-code, an ORCID API internal identifier
for a work}
}
\description{
The ORCID API has the concept of a \code{work} with its
own dedicated API endpoint. This function calls
that endpoint and returns a single data.frame
row representing that \code{work}.
}
\details{
Additional functionality is available if the
\code{work} record has a bibtex entry. Using the \code{rbibutils}
package, one can convert the citation to bibtex, latex,
citations of various styles, and even HTML.
}
\examples{
orcid = '0000-0001-5643-4068'
put_code = 89488980

df = orcid_work_detail(orcid, put_code)
df
colnames(df)

# the bibtex entry, if available,
# can be used for additional manipulation
if(require('rbibutils')) {
   rref = rbibutils::charToBib(df$citation_value)
   print(rref)

   # bibtex
   print(rref, style = "Bibtex")

   # simple citation
   print(rref, style = "citation")

   # html
   print(rref, style = "html")

   # latex
   print(rref, style = "latex")

   # r citation
   print(rref, style = "R")
}

}
\seealso{
Other orcid: 
\code{\link{.extract_simple_details_from_work_record}()},
\code{\link{orcid_works_json}()},
\code{\link{works_from_orcid}()}

Other bibliography: 
\code{\link{.extract_simple_details_from_work_record}()},
\code{\link{orcid_works_json}()},
\code{\link{works_from_orcid}()}
}
\concept{bibliography}
\concept{orcid}
