% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_work_resource.R
\name{orcid_works_json}
\alias{orcid_works_json}
\title{Fetch raw JSON for ORCID works}
\usage{
orcid_works_json(orcid)
}
\description{
Fetch raw JSON for ORCID works
}
\examples{
json_response = orcid_works_json('0000-0003-3865-7844')

str(json_response)

}
\seealso{
Other orcid: 
\code{\link{.extract_simple_details_from_work_record}()},
\code{\link{orcid_work_detail}()},
\code{\link{works_from_orcid}()}

Other bibliography: 
\code{\link{.extract_simple_details_from_work_record}()},
\code{\link{orcid_work_detail}()},
\code{\link{works_from_orcid}()}
}
\concept{bibliography}
\concept{orcid}
