% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_expertise_graph.R
\name{person_expertise_graph}
\alias{person_expertise_graph}
\title{create a graph of people and their expertise}
\usage{
person_expertise_graph(person_expertise)
}
\arguments{
\item{person_expertise}{a \code{data.frame} with people in column 1
and expertise in column 2. Other columns are ignored}
}
\value{
an \code{igraph} instance
}
\description{
create a graph of people and their expertise
}
\examples{

dssrc = dssrc_dataframe()

person_expertise = dssrc |> 
  dplyr::select(fullName, `expertise-all`) |> 
  dplyr::arrange(fullName, `expertise-all`) |> 
  dplyr::mutate(`expertise-all`=tolower(`expertise-all`)) |> 
  dplyr::mutate(`expertise-all` = gsub('\\\\(.*\\\\)', '', `expertise-all`)) |>
  tidyr::separate_longer_delim(`expertise-all`, ', ') |> 
  unique()

g = person_expertise_graph(person_expertise)

if(require(visNetwork)) {
  visNetwork::visIgraph(h,layout='layout_with_graphopt',charge=0.25)
}

}
\author{
seandavi@gmail.com
}
\concept{visualization, graph}
