% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_get}
\alias{safe_get}
\title{Get a value from an object with names}
\usage{
safe_get(l, key, default = NA)
}
\arguments{
\item{l}{an object with names}

\item{key}{the name that you want to retrieve}

\item{default}{the default value to return if the
key does not exist}
}
\description{
Get a value from an object with names
}
\examples{
l = list(a=1,b=2,c=3)
safe_get(l,'a')
safe_get(l, 'not_present')

}
\seealso{
Other utilities: 
\code{\link{has_key}()}
}
\concept{utilities}
\keyword{Internal}
