% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_work_resource.R
\name{works_from_orcid}
\alias{works_from_orcid}
\title{Works from a single ORCID}
\usage{
works_from_orcid(orcid)
}
\arguments{
\item{orcid}{character(1) an orcid is required}
}
\value{
a dataframe (see example)
}
\description{
Fetch a (simplified) set of works (publications, etc)
from an ORCID.
}
\examples{
orcid = '0000-0001-5643-4068'

df = works_from_orcid(orcid)
dim(df)
head(df)
summary(df)

}
\seealso{
Other orcid: 
\code{\link{.extract_simple_details_from_work_record}()},
\code{\link{orcid_work_detail}()},
\code{\link{orcid_works_json}()}

Other bibliography: 
\code{\link{.extract_simple_details_from_work_record}()},
\code{\link{orcid_work_detail}()},
\code{\link{orcid_works_json}()}
}
\concept{bibliography}
\concept{orcid}
